Create standard output filename from outlier diagnostics (#12)
Quick fix to standardize output filenames for outlier diagnostics to simplify
access from `pynnmap-interna`l's CLI command post-outliers. Previously,
all three diagnostics had different variable names for the output file that
need to be read in such that they can be posted to a database.

Additionally, CI tests are now disabled due to lack of setup of R on Github.