Type hinting for nn_finder.py module and associated functions (#13)
The `nn_finder.py` module does a lot of data extraction and passing with somewhat complicated data structures. This PR provides type hinting for all functions and classes within `nn_finder.py` and other functions or classes used from this module.
 
Additionally, this PR fixes an issue when a single raster path is used for more than one covariate. For example, a imagery band like `TC1` could be used in a model both for the current year as well as the previous year (model covariates `TC1` and `TC1_PREV`). Data structures within `extract_footprints` were creating a dictionary from path to a single model variable name and this has been changed to be a dictionary from path to a set of model variable by year tuples.