Calculate cross validation accuracy assessment from stored pixel file (#17)
Historically, for cross validation, plot ID values used in modeling are
captured and then pixel-level values at plot footprints are extracted
based on covariates in the model configuration.  This can be a long
process although it remains necessary when obtaining new targets for
which the user has no previously calculated covariate information.

However, for cross validation, a user can create a CSV file that stores
pixel values for all covariates for all model plots and use this
information instead.  This PR creates a new `NNFinder` derived class
(`StoredPixelNNFinder`) that simply reads in the CSV file and creates
covariate information for the model plots.  The cross-validate CLI
program looks for the presence of the `environmental_pixel_file` tag in
the configuration and, if present, uses this class, otherwise defaults
to the `PixelNNFinder` which does the extractions on the fly.

- Add optional `environmental_pixel_file` configuration option
- Calculate cross-validation from stored pixel-level files
- Pull out `run_cross_validate` as a separate function