Skip pixels with NoData in PixelNNFinder and StoredPixelNNFinder
As a result of creating the StoredPixelNNFinder and comparing it against
the PixelNNFinder, we discovered that both finders were not handling
NoData pixels correctly.  For StoredPixelNNFinder, it was a simple
matter of using `.dropna` to remove any rows with NoData.  For
PixelNNFinder, we first needed to read the rasters in with `masked=True`
and then exclude any pixels where any covariate is a masked value.

This commit also includes two unrelated small fixes that should have
been part of #17.